Added IpedsIcPy to InstitutionBuilder

+++ db/migrate/20170122220228_create_ipeds_ic_ays.rb
@@ -3,9 +3,9 @@ class CreateIpedsIcAys < ActiveRecord::Migration
     create_table :ipeds_ic_ays do |t|
       # Used in the building of DataCsv
       t.string :cross, null: false
-      t.integer :tuition_in_state
-      t.integer :tuition_out_of_state
-      t.integer :books
+      t.float :tuition_in_state
+      t.float :tuition_out_of_state
+      t.float :books
 
       # Not used in building DataCsv, but used in exporting source csv
       t.string :xtuit1
